chore: finalize language statusbar API
