+++ packages/types/vscode/typings/vscode.d.ts
@@ -2657,25 +2657,25 @@ declare module 'vscode' {
   }
 
   /**
-	 * Represents the state of a {@link Terminal}.
-	 */
-	export interface TerminalState {
-		/**
-		 * Whether the {@link Terminal} has been interacted with. Interaction means that the
-		 * terminal has sent data to the process which depending on the terminal's _mode_. By
-		 * default input is sent when a key is pressed or when a command or extension sends text,
-		 * but based on the terminal's mode it can also happen on:
-		 *
-		 * - a pointer click event
-		 * - a pointer scroll event
-		 * - a pointer move event
-		 * - terminal focus in/out
-		 *
-		 * For more information on events that can send data see "DEC Private Mode Set (DECSET)" on
-		 * https://invisible-island.net/xterm/ctlseqs/ctlseqs.html
-		 */
-		readonly isInteractedWith: boolean;
-	}
+   * Represents the state of a {@link Terminal}.
+   */
+  export interface TerminalState {
+    /**
+     * Whether the {@link Terminal} has been interacted with. Interaction means that the
+     * terminal has sent data to the process which depending on the terminal's _mode_. By
+     * default input is sent when a key is pressed or when a command or extension sends text,
+     * but based on the terminal's mode it can also happen on:
+     *
+     * - a pointer click event
+     * - a pointer scroll event
+     * - a pointer move event
+     * - terminal focus in/out
+     *
+     * For more information on events that can send data see "DEC Private Mode Set (DECSET)" on
+     * https://invisible-island.net/xterm/ctlseqs/ctlseqs.html
+     */
+    readonly isInteractedWith: boolean;
+  }
 
   //#region EnvironmentVariable
 
@@ -3842,6 +3842,83 @@ declare module 'vscode' {
     provideInlineValues(document: TextDocument, viewPort: Range, context: InlineValueContext, token: CancellationToken): ProviderResult<InlineValue[]>;
   }
   //#endregion Inline Values
+
+  /**
+   * Represents the severity of a language status item.
+   */
+  export enum LanguageStatusSeverity {
+    Information = 0,
+    Warning = 1,
+    Error = 2
+  }
+
+  /**
+   * A language status item is the preferred way to present language status reports for the active text editors,
+   * such as selected linter or notifying about a configuration problem.
+   */
+  export interface LanguageStatusItem {
+
+    /**
+     * The identifier of this item.
+     */
+    readonly id: string;
+
+    /**
+     * The short name of this item, like 'Java Language Status', etc.
+     */
+    name: string | undefined;
+
+    /**
+     * A {@link DocumentSelector selector} that defines for what editors
+     * this item shows.
+     */
+    selector: DocumentSelector;
+
+    /**
+     * The severity of this item.
+     *
+     * Defaults to {@link LanguageStatusSeverity.Information information}. You can use this property to
+     * signal to users that there is a problem that needs attention, like a missing executable or an
+     * invalid configuration.
+     */
+    severity: LanguageStatusSeverity;
+
+    /**
+     * The text to show for the entry. You can embed icons in the text by leveraging the syntax:
+     *
+     * `My text $(icon-name) contains icons like $(icon-name) this one.`
+     *
+     * Where the icon-name is taken from the ThemeIcon [icon set](https://code.visualstudio.com/api/references/icons-in-labels#icon-listing), e.g.
+     * `light-bulb`, `thumbsup`, `zap` etc.
+     */
+    text: string;
+
+    /**
+     * Optional, human-readable details for this item.
+     */
+    detail?: string;
+
+    /**
+     * Controls whether the item is shown as "busy". Defaults to `false`.
+     */
+    busy: boolean;
+
+    /**
+     * A {@linkcode Command command} for this item.
+     */
+    command: Command | undefined;
+
+    /**
+     * Accessibility information used when a screen reader interacts with this item
+     */
+    accessibilityInformation?: AccessibilityInformation;
+
+    /**
+     * Dispose and free associated resources.
+     */
+    dispose(): void;
+  }
+
 }
 
 /**
